requirements.txt
Django==5.0.6
django-ninja==1.1.0
pillow==10.3.0
